requirements.txt
azure-functions
pandas
numpy
scikit-learn
scipy
joblib
fastapi
uvicorn
streamlit
requests
